OfficeWS + Test angepasst

+++ src/de/patientenportal/services/OfficeWSImpl.java
@@ -1,13 +1,25 @@
 package de.patientenportal.services;
 
+import java.util.Arrays;
+import java.util.List;
+
 import javax.jws.WebService;
 import javax.transaction.Transactional;
+
+import de.patientenportal.entities.Access;
+import de.patientenportal.entities.ActiveRole;
 import de.patientenportal.entities.Office;
+import de.patientenportal.entities.User;
+import de.patientenportal.entities.exceptions.AccessException;
 import de.patientenportal.entities.exceptions.AccessorException;
+import de.patientenportal.entities.exceptions.AuthenticationException;
+import de.patientenportal.entities.exceptions.AuthorizationException;
 import de.patientenportal.entities.exceptions.InvalidParamException;
 import de.patientenportal.entities.exceptions.PersistenceException;
 import de.patientenportal.entities.response.Accessor;
+import de.patientenportal.persistence.DoctorDAO;
 import de.patientenportal.persistence.OfficeDAO;
+import de.patientenportal.persistence.UserDAO;
 
 @WebService(endpointInterface = "de.patientenportal.services.OfficeWS")
 public class OfficeWSImpl implements OfficeWS {
@@ -51,22 +63,29 @@ public class OfficeWSImpl implements OfficeWS {
 	 * 
 	 * @param accessor
 	 *            mit <code>Office</code> office
-	 * @return <code>String</code> response mit Erfolgsmeldung oder Fehler
+	 * @return <code>String</code> response mit Erfolgsmeldung
 	 * @throws AccessorException
 	 * @throws InvalidParamException
 	 * @throws PersistenceException
+	 * @throws AuthorizationException
+	 * @throws AccessException
+	 * @throws AuthenticationException
 	 */
 	@Transactional
-	public String createOffice(Accessor accessor)
-			throws AccessorException, InvalidParamException, PersistenceException {
+	public String createOffice(Accessor accessor) throws AccessorException, InvalidParamException, PersistenceException,
+			AuthenticationException, AccessException, AuthorizationException {
 		Office office = new Office();
+		String token;
 
 		try {
 			office = (Office) accessor.getObject();
+			token = (String) accessor.getToken();
 		} catch (Exception e) {
 			throw new AccessorException("Incorrect Accessor");
 		}
-
+		if (token == null) {
+			throw new InvalidParamException("No Token found");
+		}
 		if (office.getName() == null) {
 			throw new InvalidParamException("No Name found");
 		}
@@ -77,15 +96,16 @@ public class OfficeWSImpl implements OfficeWS {
 			throw new InvalidParamException("No Contact found");
 		}
 
-		else {
-			String response = null;
-			try {
-				response = OfficeDAO.createOffice(office);
-			} catch (Exception e) {
-				throw new PersistenceException("Error 404: Database not found");
-			}
-			return response;
+		List<ActiveRole> accesslist = Arrays.asList(ActiveRole.Doctor);
+		AuthenticationWSImpl.tokenRoleAccessCheck(accessor, accesslist, Access.Default);
+
+		String response = null;
+		try {
+			response = OfficeDAO.createOffice(office);
+		} catch (Exception e) {
+			throw new PersistenceException("Error 404: Database not found");
 		}
+		return response;
 	}
 
 	/**
@@ -93,35 +113,49 @@ public class OfficeWSImpl implements OfficeWS {
 	 * 
 	 * @param accessor
 	 *            mit <code>int</code> officeID
-	 * @return <code>String</code> response mit Erfolgsmeldung oder Fehler
+	 * @return <code>String</code> response mit Erfolgsmeldung
 	 * @throws PersistenceException
 	 * @throws InvalidParamException
 	 * @throws AccessorException
+	 * @throws AuthorizationException
+	 * @throws AccessException
+	 * @throws AuthenticationException
 	 */
-	// TODO Sicherstellen, dass ein Arzt nur seine eigene Praxis löschen kann?
 	@Transactional
-	public String deleteOffice(Accessor accessor)
-			throws PersistenceException, InvalidParamException, AccessorException {
+	public String deleteOffice(Accessor accessor) throws PersistenceException, InvalidParamException, AccessorException,
+			AuthenticationException, AccessException, AuthorizationException {
 		int id;
+		String token;
 
 		try {
-			id = (int) accessor.getObject();
+			id = (int) accessor.getId();
+			token = (String) accessor.getToken();
 		} catch (Exception e) {
 			throw new AccessorException("Incorrect Accessor");
 		}
 		if (id == 0) {
 			throw new InvalidParamException("No ID found");
 		}
-
-		else {
-			String response = null;
-			try {
-				response = OfficeDAO.deleteOffice(id);
-			} catch (Exception e) {
-				throw new PersistenceException("Error 404: Database not found");
-			}
-			return response;
+		if (token == null) {
+			throw new InvalidParamException("No Token found");
 		}
+
+		List<ActiveRole> accesslist = Arrays.asList(ActiveRole.Doctor);
+		AuthenticationWSImpl.tokenRoleAccessCheck(accessor, accesslist, Access.Default);
+
+		String response = null;
+		try {
+			User user = AuthenticationWSImpl.getUserByToken(token);
+			int officeId = UserDAO.getUser(user.getUserId()).getDoctor().getOffice().getOfficeID();
+			if (id == officeId) {
+				response = OfficeDAO.deleteOffice(id);
+			} else
+				throw new AccessException("You can only delete your own office!");
+		} catch (Exception e) {
+			throw new PersistenceException("Error 404: Database not found");
+		}
+		return response;
+
 	}
 
 	/*
@@ -135,23 +169,46 @@ public class OfficeWSImpl implements OfficeWS {
 	 * 
 	 * @param accessor
 	 *            mit <code>Office</code> office
-	 * @return <code>String</code> response mit Erfolgsmeldung oder Fehler
+	 * @return <code>String</code> response mit Erfolgsmeldung
+	 * @throws TokenException
+	 * @throws InvalidParamException
 	 * @throws AccessorException
 	 * @throws PersistenceException
+	 * @throws AuthorizationException
+	 * @throws AccessException
+	 * @throws AuthenticationException
 	 */
 	@Transactional
-	public String updateOffice(Accessor accessor) throws AccessorException, PersistenceException {
+	public String updateOffice(Accessor accessor) throws PersistenceException, InvalidParamException, AccessorException,
+			AuthenticationException, AccessException, AuthorizationException {
 		Office office = new Office();
+		String token;
+		int id;
 
 		try {
 			office = (Office) accessor.getObject();
+			token = (String) accessor.getToken();
+			id = (int) accessor.getId();
 		} catch (Exception e) {
 			throw new AccessorException("Incorrect Accessor");
 		}
+		if (token == null) {
+			throw new InvalidParamException("No Token found");
+		}
+		if (id == 0) {
+			throw new InvalidParamException("No ID found");
+		}
 
+		List<ActiveRole> accesslist = Arrays.asList(ActiveRole.Doctor);
+		AuthenticationWSImpl.tokenRoleAccessCheck(accessor, accesslist, Access.Default);
 		String response = null;
 		try {
-			response = OfficeDAO.updateOffice(office);
+			int doctorId = AuthenticationWSImpl.getUserByToken(token).getDoctor().getDoctorID();
+			int officeId = DoctorDAO.getDoctor(doctorId).getOffice().getOfficeID();
+			if (id == officeId) {
+				response = OfficeDAO.updateOffice(office);
+			} else
+				throw new AccessException("You can only update your own office!");
 		} catch (Exception e) {
 			throw new PersistenceException("Error 404: Database not found");
 		}
